refactor(tui): extract background task spawners into tasks.rs
Move spawn_request, spawn_tool_execution, and spawn_model_fetch from
mod.rs into a dedicated tasks.rs module. These ~220 lines of async
orchestration are self-contained and don't interact with the event
loop's control flow.
